requirements.txt
# ── LangGraph Orchestration (CORE PIPELINE) ──────────────────────────────────
langgraph==0.2.73
# langchain + langchain-core are transitive deps of langgraph — don't pin them
# langchain-google-vertexai only needed if GOOGLE_GENAI_USE_VERTEXAI=True

# ── Causal Discovery (THE BRAIN OF AXIOM) ────────────────────────────────────
dowhy==0.11.1
causal-learn==0.1.3.8

# ── Counterfactual Generation ─────────────────────────────────────────────────
dice-ml>=0.11

# ── Fairness Metrics Library ──────────────────────────────────────────────────
fairlearn>=0.11.0

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Google Cloud ──────────────────────────────────────────────────────────────
google-genai>=1.51.0
google-cloud-storage>=2.19.0
firebase-admin>=6.5.0

# ── Data Science ──────────────────────────────────────────────────────────────
pandas>=1.4.0
numpy>=1.26.0
scikit-learn>=1.5.0
networkx>=3.4.0
scipy>=1.14.0
joblib>=1.4.0
matplotlib>=3.9.0
graphviz>=0.20.0

# ── Utilities ─────────────────────────────────────────────────────────────────
pydantic>=2.10.0
httpx>=0.28.0
aiofiles>=24.1.0

# ── PDF Report Generation ─────────────────────────────────────────────────────
reportlab>=4.2.0
fpdf2>=2.8.0
